añadida variable gloal para el control de señales

+++ rendu4/sandbox/sandbox.c
@@ -5,8 +5,15 @@
 #include <stdlib.h>
 #include <string.h>
 #include <signal.h>
+#include <sys/errno.h>
 
-void handler(int sig){ (void)sig; }
+volatile sig_atomic_t g_timeout = 0;
+
+void handler(int sig)
+{ 
+    (void)sig;
+    g_timeout = 1;
+}
 
 int sandbox(void (*f)(void), unsigned int timeout, bool verbose)
 {
@@ -32,13 +39,15 @@ int sandbox(void (*f)(void), unsigned int timeout, bool verbose)
     int ret = waitpid(pid, &status, 0);
     alarm(0); //cancela si termina antes
 
-    if(ret == -1) //timeout
+    if(ret == -1 && errno == EINTR) //timeout
     {
         kill(pid, SIGKILL); //matar hijo
         waitpid(pid, NULL, 0); //recojo zombie
         if(verbose) printf("Bad function: timed out after %u seconds\n", timeout);
         return 0;
     }
+    if(ret == -1)
+        return -1;
 
     if(WIFEXITED(status))
     {
